Restore some fling parameters
one of the bugs I'm flailing on that are nagging at me is related to skippiness when flinging at high velocity

it's either related to constants in ViewConfiguration or(/and) animation stuff that could be handled more efficiently

MAXIMUM_FLING_VELOCITY:
stock     8000
ours was 15000

px = dp * (dpi / 160)
gsf = 430 dpi
15000 dp/s * (430 / 160) = 40312.5 px/s
on a 1080x1920 device, that's 21 screen-heights per second fling
I don't think we need to worry about flings being initiated with a 21 screen-height/second
so, returning MAXIMUM_FLING_VELOCITY to 8000 (stock val)

SCROLL_FRICTION:
stock    0.011
ours was 0.015
DADDY LIKES A SMOOTH RIDE, but this seems like it could lead to scroll inertia eventually
outrunning the scrolling cache. To prevent that, let's try the stock value.

Change-Id: I098399cf6f32661020d947eb27742bfda69161dd