fb: fix on tablets notification drawer height on custom image background
The measurment bases always on the height of the complete view. Due that we set
an image as background (does not make a difference if we do it like now with an image
view or set the background resource for the complete view) the image is calculated into
the measured height as well which leads into the problem the the panel completly opens
to the bottom.

To prevent that measure only the real content and decide based of this value how much the
rubberbands should go into action. This will be only done on devices which set a minimum
height for the drawer which indicates that we need to measure the content more exactly.

Change-Id: I9772abc8a48d5e46b5bb6e6239c0b4826115ac6a
Signed-off-by: scott <scott@ubuntu>

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PanelView.java